pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "raydog"
version = "1.0.0"
dependencies = [
    "yellowdog-sdk>=9.1.0",
    "python-dotenv==1.0.1",
    "shortuuid==1.0.13",
    "ray", 
    "ray[client]"
]